requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
baidu-aip==2.2.0.0
bleach==1.5.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
cycler==0.10.0
Django==1.11.6
django-extensions==2.0.6
django-sslserver==0.20
django-werkzeug-debugger-runserver==0.3.1
enum34==1.1.6
html5lib==0.9999999
idna==2.6
ipaddress==1.0.19
mysql-connector-python==8.0.6
protobuf==3.4.0
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.8.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
shadowsocks==2.8.2
six==1.11.0
typing==3.6.4
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.12.2
